Move fragments from AboutActivity to ConnectionListActivity(MainActivity)
