+++ OpenRedmine/src/main/AndroidManifest.xml
@@ -53,8 +53,6 @@
 				  android:configChanges="orientation|screenSize|keyboardHidden"></activity>
 		<activity android:name=".activity.KanbanActivity"
 				  android:configChanges="orientation|screenSize|keyboardHidden"></activity>
-		<activity android:name=".activity.AboutActivity"
-				  android:configChanges="orientation|screenSize|keyboardHidden"></activity>
 		<provider
 			android:name=".provider.Attachment"
 			android:authorities="jp.redmine.redmineclient.attachment"
